visualizer class split off from main
